+++ src-docs/src/views/datagrid/datagrid_styling_example.js
@@ -56,7 +56,7 @@ const gridSnippet = `<EuiDataGrid
   ]}
   // This can work as a shape.
   toolbarVisibility={{
-    showStyleSelector: false,
+    showDisplaySelector: false,
     showSortSelector: false,
     showFullScreenSelector: false,
     // showColumnSelector also takes an object, check the prop docs.
@@ -86,7 +86,7 @@ const gridSnippet = `<EuiDataGrid
     stripes: true,
     rowHover: 'highlight',
     header: 'shade',
-    // If showStyleSelector={true} from toolbarVisibility, these last two will be superceded by what the user decides.
+    // If showDisplaySelector.allowDensity={true} from toolbarVisibility, fontSize and cellPadding will be superceded by what the user decides.
     fontSize: 'm',
     cellPadding: 'm',
     footer: 'overline'
@@ -197,9 +197,10 @@ export const DataGridStylingExample = {
             individual buttons within.
           </p>
           <p>
-            With the default settings, the <EuiCode>showStyleSelector</EuiCode>{' '}
-            setting in <EuiCode>toolbarVisibility</EuiCode> means the user has
-            the ability to override the padding and font size passed into{' '}
+            With the default settings, the{' '}
+            <EuiCode>showDisplaySelector.allowDensity</EuiCode> setting in{' '}
+            <EuiCode>toolbarVisibility</EuiCode> means the user has the ability
+            to override the padding and font size passed into{' '}
             <EuiCode>gridStyle</EuiCode> by the engineer. The font size
             overriding only works with text or elements that can inherit the
             parent font size or elements that use units relative to the parent